chore: tests - implement convertMessagesToTurns function for message processing
